extract repost header to its own component
