+++ components/Note/Note.js
@@ -1,4 +1,4 @@
-import { Box, Group, Stack, Text } from "@mantine/core";
+import { Box, Group, Stack } from "@mantine/core";
 import React, { useMemo } from "react";
 import NoteTags from "../NoteTags/NoteTags";
 import NoteHeader from "../NoteHeader/NoteHeader";
@@ -9,16 +9,10 @@ import { useEvent } from "../../ndk/NDKEventProvider";
 import NoteActionRow from "../NoteActionRow/NoteActionRow";
 import { Route } from "enums";
 import NoteContent from "../NoteContent/NoteContent";
-import { RepostIcon, VerifiedIcon } from "icons/StemstrIcon";
-import useNip05 from "ndk/hooks/useNip05";
-import { Nip05Status } from "store/Nip05";
+import NoteRepostHeader from "components/NoteHeader/NoteRepostHeader";
 
 const Note = ({ type }) => {
   const { event, repostedBy } = useEvent();
-  const repostedByNip05Status = useNip05(
-    repostedBy?.hexpubkey(),
-    repostedBy?.profile?.nip05
-  );
   const { classes } = useStyles();
   const router = useRouter();
   const downloadUrl = useMemo(() => {
@@ -36,23 +30,7 @@ const Note = ({ type }) => {
 
   return (
     <Stack onClick={handleClick} sx={{ cursor: "pointer" }}>
-      {repostedBy && (
-        <Group spacing={0} c="white" fz="sm" lh="normal">
-          <RepostIcon />
-          <Text ml={8} fw="bold" span>
-            @{repostedBy.profile.name}
-          </Text>
-          {repostedByNip05Status === Nip05Status.Valid && (
-            <>
-              &nbsp;
-              <VerifiedIcon width={14} height={14} />
-            </>
-          )}
-          <Text c="gray.3" span>
-            &nbsp;reposted
-          </Text>
-        </Group>
-      )}
+      {repostedBy && <NoteRepostHeader />}
       <NoteHeader downloadUrl={downloadUrl} />
       <Group noWrap>
         {type === "parent" && (
